ocr_lightning/requirements.txt
torch
torchvision
pytorch-lightning
Pillow
# opencv-python
# tensorboard
